requirements.txt
flask>=2.0.0
gunicorn>=21.0.0
yt-dlp>=2023.0.0
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
essentia-tensorflow
